Draft making Signer contract more general
